miniProject04/pyproject.toml
[project]
name = "miniproject04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.3.0",
    "langchain-openai>=1.1.12",
    "langfuse>=2.0.0",
    "langgraph>=1.1.6",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
